my_service_fe/requirements.txt
Flask == 0.12.0
itsdangerous == 0.24
Jinja2 == 2.9.5
MarkupSafe == 0.23
Werkzeug == 0.11.15
grpcio-tools == 1.1.3

